backend/requirements.txt
aniso8601==10.0.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Admin==1.6.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pytz==2025.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
PyMySQL==1.1.0
python-dotenv
PyJWT==2.8.0
requests>=2.31.0

pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
pytest